requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
html5lib>=1.1
